[FIX, ADD][main, outputs] Fix calculation and check consistency of the results (output folder)

+++ df_types/InputDataPol.py
@@ -5,7 +5,7 @@ class InputDataPol_schema(pandera.SchemaModel):
     id_policy: pandera.typing.Index[int]
     id_pool: pandera.typing.Series[int]
     age: pandera.typing.Series[int]
-    math_res: pandera.typing.Series[int]
+    math_res: pandera.typing.Series[float]
 
 
 InputDataPol = pandera.typing.DataFrame[InputDataPol_schema]
